LayoutTests/fast/profiler/anonymous-event-handler.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

console.profile("Anonymous event handler");
function testStart()
{
    var buttonWithAnonymousHandler = document.getElementById("buttonWithAnAnonymousEventHandler")
    buttonWithAnonymousHandler.onclick = function () {
        insertNewText();
    }
    buttonWithAnonymousHandler.click();

    endTest();
}

</script>
</head>

<body onload="testStart()">
This page has an anonymous event handler.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  It should show the anonymous function at the same level as the onload
handler.
<input type="button" id="buttonWithAnAnonymousEventHandler" value="Button with an anonymous event handler">
<div id="output"></div>
</body>
</html>
